FZ-153 allow partial deletion

+++ app/Services/Client/DeleteClientService.php
@@ -2,32 +2,49 @@
 
 namespace App\Services\Client;
 
-use App\Exceptions\ApiException;
-use App\Exceptions\ExceptionFactory;
 use App\Models\Facility;
 use App\Models\Member;
 use Illuminate\Support\Facades\DB;
 
 class DeleteClientService
 {
-    public function deduplicate(Facility $facility, Member $member, string $duplicateOf)
-    {
+    public function deduplicate(
+        Facility $facility,
+        Member $member,
+        string $duplicateOf,
+    ): array {
+        // todo: replace in groups and meetings
+        return [];
     }
 
-    /** @throws ApiException */
-    public function delete(Member $member): void
-    {
+    /**
+     * @param Member $member
+     * @return array{clientDeleted: true, memberDeleted: bool, userDeleted: bool}
+     */
+    public function delete(
+        Member $member,
+    ): array {
         $client = $member->client;
         $user = $member->user;
-        if ($user->members()->whereNot('id', $member->id)->exists()) {
-            //todo another exception?
-            ExceptionFactory::userNotManagedByFacility()->throw();
-        }
-        DB::transaction(function () use ($member, $client, $user) {
-            $member->delete();
+        $deleteMember = $member->staff_member_id === null && $member->facility_admin_grant_id === null;
+        $deleteUser = $deleteMember
+            && $user->managed_by_facility_id === $member->facility_id
+            && !$user->members()->whereNot('id', $member->id)->exists();
+
+        DB::transaction(function () use ($member, $client, $user, $deleteMember, $deleteUser) {
+            // todo: delete from groups or existence of groups block deleting
+            if ($deleteMember) {
+                $member->delete();
+            } else {
+                $member->client_id = null;
+                $member->save();
+            }
             $client->values()->delete();
             $client->delete();
-            $user->delete();
+            if ($deleteUser) {
+                $user->delete();
+            }
         });
+        return ['clientDeleted' => true, 'memberDeleted' => $deleteMember, 'userDeleted' => $deleteUser];
     }
 }
